vscode/extensions/snowflake.snowflake-vsc-1.4.3/snowpark-python-template/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snowpark_example_project"
version = "0.1.0"
dependencies = [
  "snowflake-snowpark-python",
  "tomli",
  "toml",
  "snowflake-connector-python[pandas]"
]
